Add Slack notifications for CI failure monitoring
- Create standalone post_ci_failures_to_slack.py script to send CI failure alerts
- Target channel C09HCG2HM1T with @dougyster mentions when failures occur
- Filter for scheduled runs only with ≥2 consecutive failures
- Use threaded messages: parent shows workflow summary, thread shows details
- Add dummy JSON data to ci-failure-monitor.yml for Slack bot testing
- Remove unused alert_threshold parameter from ci_failures_analysis.py
- Add note that recent runs are shown left to right in failure breakdown
